new module (api_response.py) for ApiResponse
